% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{.build_request}
\alias{.build_request}
\title{Build a signed httr2 request for a given endpoint and HTTP method.}
\usage{
.build_request(endpoint, method, creds)
}
\description{
Build a signed httr2 request for a given endpoint and HTTP method.
}
\keyword{internal}
