% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{get_event}
\alias{get_event}
\title{Get a single event by ticker}
\usage{
get_event(event_ticker, creds = NULL)
}
\arguments{
\item{event_ticker}{\code{character(1)}. The event ticker.}

\item{creds}{A \code{kalshi_credentials} object or \code{NULL}.}
}
\value{
A one-row \code{tibble} with event metadata.
}
\description{
Get a single event by ticker
}
\examples{
\dontrun{
get_event("KXHIGHNY-23NOV07")
}

}
