% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{get_market}
\alias{get_market}
\title{Get a single market by ticker}
\usage{
get_market(market_ticker, creds = NULL)
}
\arguments{
\item{market_ticker}{\code{character(1)}. The market ticker.}

\item{creds}{A \code{kalshi_credentials} object or \code{NULL}.}
}
\value{
A one-row \code{tibble} with market details.
}
\description{
Get a single market by ticker
}
\examples{
\dontrun{
get_market("KXHIGHNY-23NOV07-T60")
}

}
