% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{kalshi_get_auth}
\alias{kalshi_get_auth}
\title{Retrieve the active global credentials}
\usage{
kalshi_get_auth()
}
\value{
A \code{kalshi_credentials} object.
}
\description{
Internal helper used by API functions to find credentials. Looks for
credentials set via \code{\link[=kalshi_set_auth]{kalshi_set_auth()}}. Raises a helpful error if none
are found.
}
\keyword{internal}
