% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{kalshi_post}
\alias{kalshi_post}
\title{Perform an authenticated POST request with a JSON body}
\usage{
kalshi_post(endpoint, body = list(), creds = NULL)
}
\arguments{
\item{endpoint}{Character. API path.}

\item{body}{Named list. Will be serialised to JSON.}

\item{creds}{Credentials object or NULL.}
}
\value{
Parsed JSON response as an R list.
}
\description{
Perform an authenticated POST request with a JSON body
}
\keyword{internal}
