% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markets.R
\name{parse_ob_side}
\alias{parse_ob_side}
\title{Parse one side of the orderbook into a tibble}
\usage{
parse_ob_side(levels, side)
}
\arguments{
\item{levels}{A list of (price, quantity) pairs.}

\item{side}{\code{character(1)}. \code{"yes"} or \code{"no"}.}
}
\description{
Parse one side of the orderbook into a tibble
}
\keyword{internal}
